+++ src/test/kotlin/in/kkkev/jjidea/contract/StubCommandExecutor.kt
@@ -206,6 +206,9 @@ class StubCommandExecutor(private val stub: JjStub) : CommandExecutor {
     override fun resolveList(revision: Revision): CommandExecutor.CommandResult =
         TODO("Not needed for integration tests")
 
+    override fun resolve(paths: List<String>, tool: String, revision: Revision): CommandExecutor.CommandResult =
+        TODO("Not needed for integration tests")
+
     override fun rebase(
         revisions: List<Revision>,
         destinations: List<Revision>,